+++ src/database/sqls/users_table.sql
@@ -3,6 +3,8 @@ CREATE TABLE IF NOT EXISTS users (
     username VARCHAR(50) UNIQUE NOT NULL,
     password TEXT NOT NULL,
 
+    profile_image TEXT NOT NULL,
+
     online BOOLEAN DEFAULT false NOT NULL,
     last_online TIMESTAMP NOT NULL,
     created_at TIMESTAMP NOT NULL